Rename test file, update rulesHelper functions, add unit tests, remove unnecessary unit tests from rules model
